Dodanie edycji cen pokoi, poprawki.

+++ Models/RoomType.cs
@@ -5,7 +5,7 @@
 		public int Id { get; set; }
 		public string Name { get; set; }
 		public int Beds { get; set; }
-		public decimal PricePerNight { get; set; }
+		public double PricePerNight { get; set; }
 
 		public ICollection<Room> Rooms { get; set; }
 	}
